requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
groq
openai
python-multipart
sqlalchemy
tenacity